Show all supported board types in board manager
