+++ src/views/tsconfig.json
@@ -7,7 +7,7 @@
         "outDir": "out",
         "alwaysStrict": true,
         "sourceMap": true,
-        "rootDir": "."
+        "rootDirs": [".", "../common/sharedUtilities"]
     },
     "exclude": [
         "node_modules"
